pyproject.toml
[project]
name = "external-aimbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arguably>=1.3.0",
    "lap>=0.5.12",
    "libevdev>=0.11",
    "pillow>=11.2.1",
    "ultralytics>=8.3.134",
]
